+++ src/main/java/com/farao_community/farao/gridcapa/export/adapter/FtpClientAdapter.java
@@ -6,7 +6,8 @@
  */
 package com.farao_community.farao.gridcapa.export.adapter;
 
-import com.farao_community.farao.gridcapa.export.configProperties.FtpConfigurationProperties;
+import com.farao_community.farao.gridcapa.export.config_properties.FtpConfigurationProperties;
+import com.farao_community.farao.gridcapa.export.exception.ClientAdapterException;
 import org.apache.commons.net.ftp.FTP;
 import org.apache.commons.net.ftp.FTPClient;
 import org.apache.commons.net.ftp.FTPReply;
@@ -20,6 +21,7 @@ import java.io.InputStream;
 
 /**
  * @author Mohamed BenRejeb {@literal <mohamed.ben-rejeb at rte-france.com>}
+ * @author Oualid Aloui {@literal <oualid.aloui at rte-france.com>}
  */
 
 @Component
@@ -28,36 +30,42 @@ public class FtpClientAdapter implements ClientAdapter {
     private static final Logger LOGGER = LoggerFactory.getLogger(FtpClientAdapter.class);
 
     private final FtpConfigurationProperties ftpConfigurationProperties;
-    private final FTPClient ftp = new FTPClient(); // NOSONAR
 
     public FtpClientAdapter(FtpConfigurationProperties ftpConfigurationProperties) {
         this.ftpConfigurationProperties = ftpConfigurationProperties;
     }
 
-    public void open() throws IOException {
-        ftp.connect(ftpConfigurationProperties.getHost(), ftpConfigurationProperties.getPort());
-        int reply = ftp.getReplyCode();
-        if (!FTPReply.isPositiveCompletion(reply)) {
+    public void upload(String fileName, InputStream inputStream) throws ClientAdapterException {
+        try {
+            FTPClient ftp = new FTPClient(); // NOSONAR
+            LOGGER.info("Attempt to connect to FTP server");
+            ftp.connect(ftpConfigurationProperties.getHost(), ftpConfigurationProperties.getPort());
+
+            int reply = ftp.getReplyCode();
+            if (!FTPReply.isPositiveCompletion(reply)) {
+                ftp.disconnect();
+                throw new IOException("Exception in connecting to FTP Server");
+            }
+            ftp.login(ftpConfigurationProperties.getAccessKey(), ftpConfigurationProperties.getSecretKey());
+            LOGGER.info("Connection established");
+
+            // if ftp working dir is /home/farao/upload and you want to upload files under /home/farao/upload/outputs, then the remote relative destination dir should be 'outputs', FTPClient will append it itself
+            ftp.changeWorkingDirectory(ftpConfigurationProperties.getRemoteRelativeDestinationDirectory());
+            ftp.enterLocalPassiveMode();
+            ftp.setFileType(FTP.BINARY_FILE_TYPE);  // required because ASCII is the default file type, otherwise zip will be corrupted
+            LOGGER.info("Attempt to copy {} file to FTP server", fileName);
+            boolean successFlag = ftp.storeFile(fileName, inputStream);
+            if (successFlag) {
+                LOGGER.info("File {} copied successfully to FTP server", fileName);
+            } else {
+                LOGGER.error("File {} couldn't be copied successfully to FTP server", fileName);
+            }
+
             ftp.disconnect();
-            throw new IOException("Exception in connecting to FTP Server");
+            LOGGER.info("Connection closed");
+        } catch (IOException e) {
+            LOGGER.error("Fail during upload");
+            throw new ClientAdapterException(e);
         }
-        ftp.login(ftpConfigurationProperties.getAccessKey(), ftpConfigurationProperties.getSecretKey());
-    }
-
-    public void upload(String fileName, InputStream inputStream) throws IOException {
-        // if ftp working dir is /home/farao/upload and you want to upload files under /home/farao/upload/outputs, then the remote relative destination dir should be 'outputs', FTPClient will append it itself
-        ftp.changeWorkingDirectory(ftpConfigurationProperties.getRemoteRelativeDestinationDirectory());
-        ftp.enterLocalPassiveMode();
-        ftp.setFileType(FTP.BINARY_FILE_TYPE);  // required because ASCII is the default file type, otherwise zip will be corrupted
-        boolean successFlag = ftp.storeFile(fileName, inputStream);
-        if (successFlag) {
-            LOGGER.info("File {} copied successfully to FTP server", fileName);
-        } else {
-            LOGGER.error("File {} couldn't be copied successfully to FTP server", fileName);
-        }
-    }
-
-    public void close() throws IOException {
-        ftp.disconnect();
     }
 }
