+++ src/main/java/com/farao_community/farao/gridcapa/export/adapter/SftpClientAdapter.java
@@ -6,7 +6,8 @@
  */
 package com.farao_community.farao.gridcapa.export.adapter;
 
-import com.farao_community.farao.gridcapa.export.configProperties.SftpConfigurationProperties;
+import com.farao_community.farao.gridcapa.export.config_properties.SftpConfigurationProperties;
+import com.farao_community.farao.gridcapa.export.exception.ClientAdapterException;
 import com.jcraft.jsch.ChannelSftp;
 import com.jcraft.jsch.JSch;
 import com.jcraft.jsch.JSchException;
@@ -31,31 +32,31 @@ public class SftpClientAdapter implements ClientAdapter {
 
     private final SftpConfigurationProperties sftpConfigurationProperties;
 
-    private Session jschSession;
-
     public SftpClientAdapter(SftpConfigurationProperties sftpConfigurationProperties) {
         this.sftpConfigurationProperties = sftpConfigurationProperties;
     }
 
-    public void open() throws JSchException {
-        JSch jsch = new JSch();
-        jschSession = jsch.getSession(sftpConfigurationProperties.getAccessKey(), sftpConfigurationProperties.getHost(), sftpConfigurationProperties.getPort());
-        jschSession.setPassword(sftpConfigurationProperties.getSecretKey());
-        jschSession.connect();
-    }
+    public void upload(String fileName, InputStream inputStream) throws ClientAdapterException {
+        try {
+            JSch jsch = new JSch();
+            Session jschSession = jsch.getSession(sftpConfigurationProperties.getAccessKey(), sftpConfigurationProperties.getHost(), sftpConfigurationProperties.getPort());
+            jschSession.setPassword(sftpConfigurationProperties.getSecretKey());
+            jschSession.connect();
 
-    public void upload(String fileName, InputStream inputStream) throws JSchException, SftpException {
-        ChannelSftp channelSftp = (ChannelSftp) jschSession.openChannel("sftp");
-        LOGGER.info("Attempt to connect to SFTP channel");
-        channelSftp.connect();
-        LOGGER.info("Connection established");
-        LOGGER.info("Attempt to copy {} file to SFTP server", fileName);
-        channelSftp.put(inputStream, sftpConfigurationProperties.getRemoteRelativeDestinationDirectory() + fileName);
-        LOGGER.info("File {} copied successfully to SFTP server", fileName);
-        channelSftp.disconnect();
-    }
+            ChannelSftp channelSftp = (ChannelSftp) jschSession.openChannel("sftp");
+            LOGGER.info("Attempt to connect to SFTP channel");
+            channelSftp.connect();
+            LOGGER.info("Connection established");
+            LOGGER.info("Attempt to copy {} file to SFTP server", fileName);
+            channelSftp.put(inputStream, sftpConfigurationProperties.getRemoteRelativeDestinationDirectory() + fileName);
+            LOGGER.info("File {} copied successfully to SFTP server", fileName);
 
-    public void close() {
-        jschSession.disconnect();
+            channelSftp.disconnect();
+            jschSession.disconnect();
+            LOGGER.info("Connection closed");
+        } catch (JSchException | SftpException e) {
+            LOGGER.error("Fail during upload");
+            throw new ClientAdapterException(e);
+        }
     }
 }
