+++ src/main/java/com/farao_community/farao/gridcapa/export/config_properties/FtpConfigurationProperties.java
@@ -4,26 +4,25 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.gridcapa.export.configProperties;
+package com.farao_community.farao.gridcapa.export.config_properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.ConstructorBinding;
 
 /**
- * @author Oualid Aloui {@literal <oualid.aloui at rte-france.com>}
+ * @author Mohamed Benrejeb {@literal <mohamed.ben-rejeb at rte-france.com>}
  */
 
 @ConstructorBinding
-@ConfigurationProperties("sftp")
-public class SftpConfigurationProperties {
+@ConfigurationProperties("ftp")
+public class FtpConfigurationProperties {
+    private String host;
+    private int port;
+    private String accessKey;
+    private String secretKey;
+    private String remoteRelativeDestinationDirectory;
 
-    private final String host;
-    private final int port;
-    private final String accessKey;
-    private final String secretKey;
-    private final String remoteRelativeDestinationDirectory;
-
-    public SftpConfigurationProperties(String host, int port, String accessKey, String secretKey, String remoteRelativeDestinationDirectory) {
+    public FtpConfigurationProperties(String host, int port, String accessKey, String secretKey, String remoteRelativeDestinationDirectory) {
         this.host = host;
         this.port = port;
         this.accessKey = accessKey;
